+++ src/providers/App/AppProvider.tsx
@@ -1,15 +1,17 @@
 'use client';
 
 import { ReactNode, useReducer, useMemo, Context, createContext, useContext } from 'react';
+import { useAccount } from 'wagmi';
 import useSafeContracts from '../../hooks/safe/useSafeContracts';
 import { FractalStore, StoreAction } from '../../types';
+import { combinedReducer, initialState } from './combinedReducer';
 import { useSafeService } from './hooks/useSafeService';
-import { initialState, useCombinedReducer } from './useCombinedReducer';
+import { useReadOnlyValues } from './useReadOnlyValues';
 export const FractalContext = createContext<FractalStore | null>(null);
 
 export const useFractal = (): FractalStore => useContext(FractalContext as Context<FractalStore>);
 export function AppProvider({ children }: { children: ReactNode }) {
-  const combinedReducer = useCombinedReducer();
+  const { address: account } = useAccount();
   // Replace individual useReducer calls with a single combined reducer
   const [state, dispatch] = useReducer(combinedReducer, initialState);
   // loads base Fractal contracts with provider into state
@@ -17,6 +19,7 @@ export function AppProvider({ children }: { children: ReactNode }) {
   // loads safe service into state;
   const safeService = useSafeService();
   // memoize fractal store
+  const { readOnlyValues } = useReadOnlyValues();
   const fractalStore: FractalStore = useMemo(() => {
     return {
       node: state.node,
@@ -25,7 +28,7 @@ export function AppProvider({ children }: { children: ReactNode }) {
       treasury: state.treasury,
       governanceContracts: state.governanceContracts,
       guardContracts: state.guardContracts,
-      readOnly: state.readOnly,
+      readOnly: readOnlyValues(state, account),
       action: {
         dispatch,
         resetDAO: async () => {
@@ -37,7 +40,7 @@ export function AppProvider({ children }: { children: ReactNode }) {
       },
       baseContracts,
     };
-  }, [state, safeService, baseContracts]);
+  }, [state, safeService, baseContracts, account, readOnlyValues]);
 
   return <FractalContext.Provider value={fractalStore}>{children}</FractalContext.Provider>;
 }
